+++ affinidi-messaging-sdk/src/messages/mod.rs
@@ -7,6 +7,7 @@ pub mod fetch;
 pub mod get;
 pub mod list;
 pub mod pack;
+pub mod problem_report;
 pub mod sending;
 pub mod unpack;